requirements.txt
streamlit>=1.31.0
langchain>=0.1.0
langchain-community>=0.0.20
langchain-openai>=0.0.5
langchain-text-splitters>=0.0.1
langchain-core>=0.1.0
openai>=1.12.0
faiss-cpu>=1.9.0
sentence-transformers>=2.3.0
requests>=2.31.0
python-dotenv>=1.0.0
pandas>=2.0.0
numpy>=1.24.0
tiktoken>=0.6.0
